[ENHANCEMENT] Export prop type for `TemplateVariableProvider` and refactor storybook decorators
