+++ ui/storybook/src/index.ts
@@ -11,4 +11,5 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+export * from './decorators';
 export * from './utils';
